Update `ipset` example
The `ipset` script looked to be designed so that it could be executed repeatedly without issues, due to the `flush` statement and `-q` parameters.
However, doing so would generate duplicate iptables entries and temporarily disable the blocklist while new IPs were being fetched.
The new script solves these issues and uses apt-get to avoid a CLI warning when executed from cron.

If this new example script is too complex I would instead suggest removing the `ipset -q flush ipsum` line and make it more clear that it's not a ready-to-deploy script.